add to check on valid id

+++ server/utils.py
@@ -1,4 +1,5 @@
 from server.datastore import chat_dao, user_dao
+from server.datastore.datastore import id_is_valid
 from server.entities.chats.dialog import Dialog
 
 
@@ -7,6 +8,8 @@ def create_dialog(user_id_1, user_id_2):
        then add this dialog to users on mongoDB"""
 
     if not user_dao.is_exists(user_id_1) or not user_dao.is_exists(user_id_2):
+        print("create_dialog: id='{id1}' or id='{id2}' are not valid or are not exists".format(id1=user_id_1,
+                                                                                               id2=user_id_2))
         return None
 
     # create and save dialog
@@ -23,10 +26,13 @@ def create_dialog(user_id_1, user_id_2):
 def delete_dialog(dialog_id):
     """Delete dialog, remove it in users, remove all msg's in it"""
 
+    if not chat_dao.dialog_is_exist(dialog_id):
+        print("delete_dialog: id='{id}' is not valid or is not exists".format(id=dialog_id))
+        return
+
     # get dialog by id
     dialog = chat_dao.get_dialog(dialog_id)
 
-    # TODO check that dialog is not null
     # delete dialog from user
     user_dao.delete_chat(dialog.user_id_1, dialog_id)
     user_dao.delete_chat(dialog.user_id_2, dialog_id)
